Add new ignore settings and Docker configurations
Added new files for .idea settings to be ignored, which includes workspace.xml, tasks.xml, dictionaries, and dataSources among others. This is to ensure that our git repository won't include any unnecessary files and to keep it clean. Also, added a Dockerfile for the purpose of creating a PostgreSQL Kubernetes operator, and updated .gitignore file based on common IntelliJ ignores and high-churn files. This will allow us to have a containerised version of our application that can be deployed more conveniently.