Add --break-system-packages flag to pip install
This commit adds --break-system-packages flag to pip3 install commands in the Dockerfile.
This change appears to be required to complete the installation of 'rsa' and 'python-magic' packages successfully, as indicated by an issue linked in an askUbuntu thread (#1465218). The exact implications of this flag are not entirely clear but it resolves the installation issue.