Update postgres image and add pgbackrest service
The postgres docker image is updated to `local` in the docker-compose file. Removal of 'always' pull policy is also included. Additionally, the pgbackrest service has been added, which provides backup capabilities for PostgreSQL. The restart policy of the pgadmin service has been also modified to 'on-failure'. New volumes for pgbackrest are introduced as well.