Update docker-compose and add pgbackrest configuration
This commit updates the docker-compose-db.yml file to modify the volume path of pgbackrest configuration and introduces a new service called network-multitool for network troubleshooting. More importantly, a new configuration file, pgbackrest.conf, has been added to the repository, which is used for managing backups and restores of the PostgreSQL database.